V-Dem dataset (version 7.1)

The V-Dem dataset (version 7.1), including external variables. This documentation only provides variable names and labels for the country-year identifiers included. The rest of the codebook is available at https://www.v-dem.net/en/data/data-version-7-1/, and in this package as vdem_codebook.

VDem_plus

Format

An object of class tbl_df (inherits from tbl, data.frame) with 17604 rows and 3615 columns.

Source

Coppedge, Michael, John Gerring, Staffan I. Lindberg, Svend-Erik Skaaning, Jan Teorell, David Altman, Michael Bernhard, M. Steven Fish, Adam Glynn, Allen Hicken, Carl Henrik Knutsen, Joshua Krusell, Anna Luhrmann, Kyle L. Marquardt, Kelly McMann, Valeriya Mechkova, Moa Olin, Pamela Paxton, Daniel Pemstein, Josefine Pernes, Constanza Sanhueza Petrarca, Johannes von Romer, Laura Saxer, Brigitte Seim, Rachel Sigman, Jeffrey Staton, Natalia Stepanova, and Steven Wilson. 2017. "V-Dem Country-Year/Country-Date Dataset v7.1" Varieties of Democracy (V-Dem) Project. Available at https://www.v-dem.net/en/data/data-version-7-1/

Identifiers

vdem_country_name

The original country name in the dataset. There are 178 distinct countries in version 7.1. According to the codebook, "a V-Dem country is a political unit enjoying at least some degree of functional and/or formal sovereignty."

vdem_country_id

Unique numerical country id for each country.

vdem_country_text_id

Unique abbreviation for each country.

year

Year, from 1900-2016.

extended_country_name

The standard country according to the Gleditsch and Ward system ofstates (Gleditsch and Ward 1999), or a standardized country name. Note that West Bank, and Gaza are two separate coding units in vdem_country_name but all map to "Palestine, State of" here. Column added by Xavier Marquez to the original V-Dem dataset.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states. This differs from the V-Dem COW code in a few instances. For details, see Gleditsch and Ward 1999. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html. Column added by Xavier Marquez to the original V-Dem dataset.

historical_date

According to the codebook, "This variable is included in the V-Dem Country Date dataset. The default date is 31st December, as in 2016-12-31, referring to the time span from 01-01 to 12-31 in a respective year. Additionally, specific changes, such as the appointment of a Head of State, are coded on the specific date within a certain year. Thus, a code can change within a year, and will be reflected in the 12-31 date."

codingstart

"V-Dem country coding starts in 1900, or from when a country first was formally sovereign (according to international law). For detailed information, please see to VDem Country Coding Units."

gapstart

"Time periods when a country does not fulfill V-Dem’s coding period criteria are not coded. The date that indicates the gap start is the last date coded before the gap. For more details about V-Dem country coding periods, please see to V-Dem Country Coding Units."

gapend

"The periods of when a country does not fulfill V-Dem’s coding period criteria are not coded. The date that indicates the gap end is the first date coded after the gap. For more details about V-Dem country coding periods, please see to V-Dem Country Coding Units."

codingend

"V-Dem country coding ends in 2016, or from when a country formally stopped being sovereign (according to international law). For detailed information , please see VDem Country Coding Units."

vdem_cown

The Correlates of War numeric country code. Note that this differs from Gleditsch and Ward's numeric country code in a few cases.

cown

The Correlates of War numeric country code (from the 2016 version of the COW state list). Note that this differs from Gleditsch and Ward's numeric country code in a few cases; it also differs from vdem_cown in one case (Germany 1949-1989), where vdem_cown incorrectly has 255 (it should be 260). Column added by Xavier Marquez to the original V-Dem dataset.

GWc

Gleditsch and Ward's alphabetic country code, from the Gleditsch and Ward list of independent states. For details, see Gleditsch and Ward 1999. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html. Column added by Xavier Marquez to the original V-Dem dataset.

region

Region. Column added by Xavier Marquez to the original V-Dem dataset.

continent

Continent. Column added by Xavier Marquez to the original V-Dem dataset.

GW_startdate

The entry date of the state into the international system (usually the date of independence, though there can be more than one), according to Gleditsch and Ward, or 1 January 1816, whichever is later. Column added by Xavier Marquez to the original V-Dem dataset.

GW_enddate

The exit date of the state from the international system (usually the date the state lost its independence, though there can be more than one), according to Gleditsch and Ward, or NA if the country is still in the international system as of the time the dataset was compiled. Column added by Xavier Marquez to the original V-Dem dataset.

microstate

Whether the state is a microstate, according to Gleditsch. His tentative list of microstates is available at http://privatewww.essex.ac.uk/~ksg/statelist.html. Column added by Xavier Marquez to the original V-Dem dataset.

lat

The rough latitude of the state. Column added by Xavier Marquez to the original V-Dem dataset.

lon

The rough longitude of the state. Column added by Xavier Marquez to the original V-Dem dataset.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Column added by Xavier Marquez to the original V-Dem dataset. You may instead wish to use one of the v2sv* columns provided in the original V-Dem dataset.

References

Coppedge, Michael, John Gerring, Staffan I. Lindberg, Svend-Erik Skaaning, Jan Teorell, David Altman, Frida Andersson, Michael Bernhard, M. Steven Fish, Adam Glynn, Allen Hicken, Carl Henrik Knutsen, Kyle L. Marquardt, Kelly McMann, Valeriya Mechkova, Pamela Paxton, Daniel Pemstein, Laura Saxer, Brigitte Seim, Rachel Sigman and Jeffrey Staton. 2017. "V-Dem Codebook v7.1" Varieties of Democracy (V-Dem) Project. Available at https://www.v-dem.net/en/data/data-version-7-1/

Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html