Extracts parts of the V-Dem dataset by section number, name pattern, or label pattern

Extracts variables in the V-Dem dataset (version 9), according to the section number in the codebook or a regular expression pattern matching the name or the label.

extract_vdem(section_number = NULL, name_pattern = NULL,
  label_pattern = NULL, include_uncertainty = TRUE,
  include_sd = FALSE, include_osp = FALSE, include_ord = FALSE,
  include_nr = FALSE, include_external = FALSE)

Arguments

section_number
A section number (or vector of section numbers) in the V-dem codebook. Must be an integer or an integer vector (e.g., 1:2, or 28).

name_pattern
A character pattern, interpreted as a regular expression, to search over the V-dem variable names, e.g., "v2x" extracts all variables including the string "v2x" in their names.

label_pattern
A character pattern, interpreted as a regular expression, to search over the V-dem labels, e.g., "corrupt" extracts all variables that mention "corrupt" in their label.

include_uncertainty
Whether to return the uncertainty measurements for the selected variables (codelow and codehigh suffixes). Default is TRUE.

include_sd
Whether to return the standard deviation for the selected variables (sd suffix). Default is FALSE.

include_osp
Whether to return variables available in the original scale (osp suffix, linearized original scale posterior prediction). Default is FALSE.

include_ord
Whether to return variables available in an ordinal scale (ord suffix, measurement model estimates of original scale value). Default is FALSE.

include_nr
Whether to return the number of coders for the variable (nr suffix). Default is FALSE.

include_external
Whether to return variables marked "external", i.e., coming from an external source rather than V-Dem (e_ prefix). Default is FALSE.

Value

A tibble with the V-Dem country identifiers, other state system identifiers, plus the desired variables.

Examples

# Main democracy indexes
extract_vdem(section_number = 1)

#> # A tibble: 26,884 x 32
#>    vdem_country_na~ vdem_country_te~ vdem_country_id  year extended_countr~
#>    <chr>            <chr>                      <dbl> <dbl> <chr>           
#>  1 Afghanistan      AFG                           36  1789 Afghanistan     
#>  2 Afghanistan      AFG                           36  1790 Afghanistan     
#>  3 Afghanistan      AFG                           36  1791 Afghanistan     
#>  4 Afghanistan      AFG                           36  1792 Afghanistan     
#>  5 Afghanistan      AFG                           36  1793 Afghanistan     
#>  6 Afghanistan      AFG                           36  1794 Afghanistan     
#>  7 Afghanistan      AFG                           36  1795 Afghanistan     
#>  8 Afghanistan      AFG                           36  1796 Afghanistan     
#>  9 Afghanistan      AFG                           36  1797 Afghanistan     
#> 10 Afghanistan      AFG                           36  1798 Afghanistan     
#> # ... with 26,874 more rows, and 27 more variables: GWn <dbl>,
#> #   historical_date <date>, project <dbl>, historical <int>, histname <chr>,
#> #   codingstart <dbl>, codingend <dbl>, codingstart_contemp <dbl>,
#> #   codingend_contemp <dbl>, codingstart_hist <dbl>, codingend_hist <dbl>,
#> #   gapstart1 <dbl>, gapstart2 <dbl>, gapstart3 <dbl>, gapend1 <dbl>,
#> #   gapend2 <dbl>, gapend3 <dbl>, vdem_cown <int>, GW_startdate <date>,
#> #   GW_enddate <date>, GWc <chr>, extended_region <chr>,
#> #   extended_continent <chr>, microstate <lgl>, lat <dbl>, lon <dbl>,
#> #   in_GW_system <lgl>

# Mid-level indexes, no measures of uncertainty
extract_vdem(section_number = 2, include_uncertainty = FALSE)

#> # A tibble: 26,884 x 58
#>    vdem_country_na~ vdem_country_te~ vdem_country_id  year extended_countr~
#>    <chr>            <chr>                      <dbl> <dbl> <chr>           
#>  1 Afghanistan      AFG                           36  1789 Afghanistan     
#>  2 Afghanistan      AFG                           36  1790 Afghanistan     
#>  3 Afghanistan      AFG                           36  1791 Afghanistan     
#>  4 Afghanistan      AFG                           36  1792 Afghanistan     
#>  5 Afghanistan      AFG                           36  1793 Afghanistan     
#>  6 Afghanistan      AFG                           36  1794 Afghanistan     
#>  7 Afghanistan      AFG                           36  1795 Afghanistan     
#>  8 Afghanistan      AFG                           36  1796 Afghanistan     
#>  9 Afghanistan      AFG                           36  1797 Afghanistan     
#> 10 Afghanistan      AFG                           36  1798 Afghanistan     
#> # ... with 26,874 more rows, and 53 more variables: GWn <dbl>,
#> #   historical_date <date>, project <dbl>, historical <int>, histname <chr>,
#> #   codingstart <dbl>, codingend <dbl>, codingstart_contemp <dbl>,
#> #   codingend_contemp <dbl>, codingstart_hist <dbl>, codingend_hist <dbl>,
#> #   gapstart1 <dbl>, gapstart2 <dbl>, gapstart3 <dbl>, gapend1 <dbl>,
#> #   gapend2 <dbl>, gapend3 <dbl>, vdem_cown <int>, v2x_polyarchy <dbl>,
#> #   v2x_libdem <dbl>, v2x_partipdem <dbl>, v2x_delibdem <dbl>,
#> #   v2x_egaldem <dbl>, v2x_api <dbl>, v2x_mpi <dbl>, v2x_freexp_altinf <dbl>,
#> #   v2x_frassoc_thick <dbl>, v2x_suffr <dbl>, v2xel_frefair <dbl>,
#> #   v2x_elecoff <dbl>, v2x_liberal <dbl>, v2xcl_rol <dbl>, v2x_jucon <dbl>,
#> #   v2xlg_legcon <dbl>, v2x_partip <dbl>, v2x_cspart <dbl>, v2xdd_dd <dbl>,
#> #   v2xel_locelec <dbl>, v2xel_regelec <dbl>, v2xdl_delib <dbl>,
#> #   v2x_egal <dbl>, v2xeg_eqprotec <dbl>, v2xeg_eqaccess <dbl>,
#> #   v2xeg_eqdr <dbl>, GW_startdate <date>, GW_enddate <date>, GWc <chr>,
#> #   extended_region <chr>, extended_continent <chr>, microstate <lgl>,
#> #   lat <dbl>, lon <dbl>, in_GW_system <lgl>

# All indexes about corruption
extract_vdem(label_pattern = "corrupt")

#> # A tibble: 26,884 x 62
#>    vdem_country_na~ vdem_country_te~ vdem_country_id  year extended_countr~
#>    <chr>            <chr>                      <dbl> <dbl> <chr>           
#>  1 Afghanistan      AFG                           36  1789 Afghanistan     
#>  2 Afghanistan      AFG                           36  1790 Afghanistan     
#>  3 Afghanistan      AFG                           36  1791 Afghanistan     
#>  4 Afghanistan      AFG                           36  1792 Afghanistan     
#>  5 Afghanistan      AFG                           36  1793 Afghanistan     
#>  6 Afghanistan      AFG                           36  1794 Afghanistan     
#>  7 Afghanistan      AFG                           36  1795 Afghanistan     
#>  8 Afghanistan      AFG                           36  1796 Afghanistan     
#>  9 Afghanistan      AFG                           36  1797 Afghanistan     
#> 10 Afghanistan      AFG                           36  1798 Afghanistan     
#> # ... with 26,874 more rows, and 57 more variables: GWn <dbl>,
#> #   historical_date <date>, project <dbl>, historical <int>, histname <chr>,
#> #   codingstart <dbl>, codingend <dbl>, codingstart_contemp <dbl>,
#> #   codingend_contemp <dbl>, codingstart_hist <dbl>, codingend_hist <dbl>,
#> #   gapstart1 <dbl>, gapstart2 <dbl>, gapstart3 <dbl>, gapend1 <dbl>,
#> #   gapend2 <dbl>, gapend3 <dbl>, vdem_cown <int>, v2exbribe <dbl>,
#> #   v2excrptps <dbl>, v2lgcrrpt <dbl>, v2jucorrdc <dbl>, v2mecorrpt <dbl>,
#> #   v3lgcrrpt <dbl>, v2xnp_regcorr <dbl>, v2x_corr <dbl>, v2x_execorr <dbl>,
#> #   v2x_pubcorr <dbl>, v2exbribe_codelow <dbl>, v2excrptps_codelow <dbl>,
#> #   v2lgcrrpt_codelow <dbl>, v2jucorrdc_codelow <dbl>,
#> #   v2mecorrpt_codelow <dbl>, v3lgcrrpt_codelow <dbl>,
#> #   v2xnp_regcorr_codelow <dbl>, v2x_corr_codelow <dbl>,
#> #   v2x_execorr_codelow <dbl>, v2x_pubcorr_codelow <dbl>,
#> #   v2exbribe_codehigh <dbl>, v2excrptps_codehigh <dbl>,
#> #   v2lgcrrpt_codehigh <dbl>, v2jucorrdc_codehigh <dbl>,
#> #   v2mecorrpt_codehigh <dbl>, v3lgcrrpt_codehigh <dbl>,
#> #   v2xnp_regcorr_codehigh <dbl>, v2x_corr_codehigh <dbl>,
#> #   v2x_execorr_codehigh <dbl>, v2x_pubcorr_codehigh <dbl>,
#> #   GW_startdate <date>, GW_enddate <date>, GWc <chr>, extended_region <chr>,
#> #   extended_continent <chr>, microstate <lgl>, lat <dbl>, lon <dbl>,
#> #   in_GW_system <lgl>

# All indexes about elections, include external indexes
extract_vdem(label_pattern = "elect", include_external = TRUE)

#> # A tibble: 26,884 x 173
#>    vdem_country_na~ vdem_country_te~ vdem_country_id  year extended_countr~
#>    <chr>            <chr>                      <dbl> <dbl> <chr>           
#>  1 Afghanistan      AFG                           36  1789 Afghanistan     
#>  2 Afghanistan      AFG                           36  1790 Afghanistan     
#>  3 Afghanistan      AFG                           36  1791 Afghanistan     
#>  4 Afghanistan      AFG                           36  1792 Afghanistan     
#>  5 Afghanistan      AFG                           36  1793 Afghanistan     
#>  6 Afghanistan      AFG                           36  1794 Afghanistan     
#>  7 Afghanistan      AFG                           36  1795 Afghanistan     
#>  8 Afghanistan      AFG                           36  1796 Afghanistan     
#>  9 Afghanistan      AFG                           36  1797 Afghanistan     
#> 10 Afghanistan      AFG                           36  1798 Afghanistan     
#> # ... with 26,874 more rows, and 168 more variables: GWn <dbl>,
#> #   historical_date <date>, project <dbl>, historical <int>, histname <chr>,
#> #   codingstart <dbl>, codingend <dbl>, codingstart_contemp <dbl>,
#> #   codingend_contemp <dbl>, codingstart_hist <dbl>, codingend_hist <dbl>,
#> #   gapstart1 <dbl>, gapstart2 <dbl>, gapstart3 <dbl>, gapend1 <dbl>,
#> #   gapend2 <dbl>, gapend3 <dbl>, vdem_cown <int>, v2xel_frefair <dbl>,
#> #   v2elpeace <dbl>, v2elffelr <dbl>, v2elffelrbin <dbl>, v2elsnlsff <dbl>,
#> #   v2eldommon <dbl>, v2ellocons <dbl>, v2ellocumul <dbl>, v2elloeldm <dbl>,
#> #   v2elloelsy <dbl>, v2elloseat <dbl>, v2ellostlg <dbl>, v2ellostsl <dbl>,
#> #   v2ellostsm <dbl>, v2ellostss <dbl>, v2ellosttm <dbl>, v2ellostts <dbl>,
#> #   v2ellovtlg <dbl>, v2ellovtsm <dbl>, v2ellovttm <dbl>, v2elparlel <dbl>,
#> #   v2elprescons <dbl>, v2elprescumul <dbl>, v2elthresh <dbl>, v2eltvrig <dbl>,
#> #   v2elvotlrg <dbl>, v2elvotsml <dbl>, v2pscnslnl <dbl>, v2exaphogp <dbl>,
#> #   v2exaphos <dbl>, v2ex_elechog <dbl>, v2ex_elechos <dbl>, v2lgelecup <dbl>,
#> #   v2lgello <dbl>, v2lginello <dbl>, v2lginelup <dbl>, v3eldirepr <dbl>,
#> #   v3eldireuc <dbl>, v3eltvriguc <dbl>, v3elupseat <dbl>, v3elupstsl <dbl>,
#> #   v3elupstsm <dbl>, v3elupvtlg <dbl>, v3elupvtsm <dbl>, v3elloelsy <dbl>,
#> #   v3elloseat <dbl>, v3ellostlg <dbl>, v3ellostsl <dbl>, v3ellostsm <dbl>,
#> #   v3ellostss <dbl>, v3ellovtlg <dbl>, v3ellovtsm <dbl>, v3elparlel <dbl>,
#> #   v3eltvrig <dbl>, v3elvotlrg <dbl>, v3elvotsml <dbl>, v3lgelecup <dbl>,
#> #   v3lgello <dbl>, v3lginello <dbl>, v3lginelup <dbl>, v3ellocelc <dbl>,
#> #   v3elsrgel <dbl>, v2x_ex_direlect <dbl>, v2xex_elecreg <dbl>,
#> #   v2xlg_elecreg <dbl>, v2x_hosabort <dbl>, v2x_hosinter <dbl>,
#> #   v2x_legabort <dbl>, v2xel_elecparl <dbl>, v2xel_elecpres <dbl>,
#> #   v2xex_elecleg <dbl>, v2elsnlfc_0 <dbl>, v2elsnlfc_1 <dbl>,
#> #   v2elsnlfc_10 <dbl>, v2elsnlfc_11 <dbl>, v2elsnlfc_12 <dbl>,
#> #   v2elsnlfc_13 <dbl>, v2elsnlfc_14 <dbl>, v2elsnlfc_15 <dbl>,
#> #   v2elsnlfc_16 <dbl>, v2elsnlfc_17 <dbl>, v2elsnlfc_18 <dbl>, ...

# All civil liberties indexes, include original scale prediction
extract_vdem(name_pattern = "^v2cl", include_osp = TRUE)

#> # A tibble: 26,884 x 223
#>    vdem_country_na~ vdem_country_te~ vdem_country_id  year extended_countr~
#>    <chr>            <chr>                      <dbl> <dbl> <chr>           
#>  1 Afghanistan      AFG                           36  1789 Afghanistan     
#>  2 Afghanistan      AFG                           36  1790 Afghanistan     
#>  3 Afghanistan      AFG                           36  1791 Afghanistan     
#>  4 Afghanistan      AFG                           36  1792 Afghanistan     
#>  5 Afghanistan      AFG                           36  1793 Afghanistan     
#>  6 Afghanistan      AFG                           36  1794 Afghanistan     
#>  7 Afghanistan      AFG                           36  1795 Afghanistan     
#>  8 Afghanistan      AFG                           36  1796 Afghanistan     
#>  9 Afghanistan      AFG                           36  1797 Afghanistan     
#> 10 Afghanistan      AFG                           36  1798 Afghanistan     
#> # ... with 26,874 more rows, and 218 more variables: GWn <dbl>,
#> #   historical_date <date>, project <dbl>, historical <int>, histname <chr>,
#> #   codingstart <dbl>, codingend <dbl>, codingstart_contemp <dbl>,
#> #   codingend_contemp <dbl>, codingstart_hist <dbl>, codingend_hist <dbl>,
#> #   gapstart1 <dbl>, gapstart2 <dbl>, gapstart3 <dbl>, gapend1 <dbl>,
#> #   gapend2 <dbl>, gapend3 <dbl>, vdem_cown <int>, v2clacfree <dbl>,
#> #   v2clrelig <dbl>, v2cltort <dbl>, v2clkill <dbl>, v2cltrnslw <dbl>,
#> #   v2clrspct <dbl>, v2clfmove <dbl>, v2cldmovem <dbl>, v2cldmovew <dbl>,
#> #   v2cldiscm <dbl>, v2cldiscw <dbl>, v2clslavem <dbl>, v2clslavef <dbl>,
#> #   v2clstown <dbl>, v2clprptym <dbl>, v2clprptyw <dbl>, v2clacjstm <dbl>,
#> #   v2clacjstw <dbl>, v2clacjust <dbl>, v2clsocgrp <dbl>, v2clrgunev <dbl>,
#> #   v2clsnlpct <dbl>, v2clgencl <dbl>, v2clpolcl <dbl>, v2clgeocl <dbl>,
#> #   v2clrgstch_0 <dbl>, v2clrgstch_1 <dbl>, v2clrgstch_10 <dbl>,
#> #   v2clrgstch_11 <dbl>, v2clrgstch_12 <dbl>, v2clrgstch_13 <dbl>,
#> #   v2clrgstch_14 <dbl>, v2clrgstch_15 <dbl>, v2clrgstch_16 <dbl>,
#> #   v2clrgstch_17 <dbl>, v2clrgstch_18 <dbl>, v2clrgstch_19 <dbl>,
#> #   v2clrgstch_2 <dbl>, v2clrgstch_20 <dbl>, v2clrgstch_21 <dbl>,
#> #   v2clrgstch_3 <dbl>, v2clrgstch_4 <dbl>, v2clrgstch_5 <dbl>,
#> #   v2clrgstch_6 <dbl>, v2clrgstch_7 <dbl>, v2clrgstch_8 <dbl>,
#> #   v2clrgstch_9 <dbl>, v2clrgwkch_0 <dbl>, v2clrgwkch_1 <dbl>,
#> #   v2clrgwkch_10 <dbl>, v2clrgwkch_11 <dbl>, v2clrgwkch_12 <dbl>,
#> #   v2clrgwkch_13 <dbl>, v2clrgwkch_14 <dbl>, v2clrgwkch_15 <dbl>,
#> #   v2clrgwkch_16 <dbl>, v2clrgwkch_17 <dbl>, v2clrgwkch_18 <dbl>,
#> #   v2clrgwkch_19 <dbl>, v2clrgwkch_2 <dbl>, v2clrgwkch_20 <dbl>,
#> #   v2clrgwkch_21 <dbl>, v2clrgwkch_3 <dbl>, v2clrgwkch_4 <dbl>,
#> #   v2clrgwkch_5 <dbl>, v2clrgwkch_6 <dbl>, v2clrgwkch_7 <dbl>,
#> #   v2clrgwkch_8 <dbl>, v2clrgwkch_9 <dbl>, v2clacfree_osp <dbl>,
#> #   v2clrelig_osp <dbl>, v2cltort_osp <dbl>, v2clkill_osp <dbl>,
#> #   v2cltrnslw_osp <dbl>, v2clrspct_osp <dbl>, v2clfmove_osp <dbl>,
#> #   v2cldmovem_osp <dbl>, v2cldmovew_osp <dbl>, v2cldiscm_osp <dbl>,
#> #   v2cldiscw_osp <dbl>, v2clslavem_osp <dbl>, v2clslavef_osp <dbl>, ...

# Combine section and label patterns
extract_vdem(section_number = 1, label_pattern = "elect", include_external = TRUE)

#> # A tibble: 26,884 x 173
#>    vdem_country_na~ vdem_country_te~ vdem_country_id  year extended_countr~
#>    <chr>            <chr>                      <dbl> <dbl> <chr>           
#>  1 Afghanistan      AFG                           36  1789 Afghanistan     
#>  2 Afghanistan      AFG                           36  1790 Afghanistan     
#>  3 Afghanistan      AFG                           36  1791 Afghanistan     
#>  4 Afghanistan      AFG                           36  1792 Afghanistan     
#>  5 Afghanistan      AFG                           36  1793 Afghanistan     
#>  6 Afghanistan      AFG                           36  1794 Afghanistan     
#>  7 Afghanistan      AFG                           36  1795 Afghanistan     
#>  8 Afghanistan      AFG                           36  1796 Afghanistan     
#>  9 Afghanistan      AFG                           36  1797 Afghanistan     
#> 10 Afghanistan      AFG                           36  1798 Afghanistan     
#> # ... with 26,874 more rows, and 168 more variables: GWn <dbl>,
#> #   historical_date <date>, project <dbl>, historical <int>, histname <chr>,
#> #   codingstart <dbl>, codingend <dbl>, codingstart_contemp <dbl>,
#> #   codingend_contemp <dbl>, codingstart_hist <dbl>, codingend_hist <dbl>,
#> #   gapstart1 <dbl>, gapstart2 <dbl>, gapstart3 <dbl>, gapend1 <dbl>,
#> #   gapend2 <dbl>, gapend3 <dbl>, vdem_cown <int>, v2xel_frefair <dbl>,
#> #   v2elpeace <dbl>, v2elffelr <dbl>, v2elffelrbin <dbl>, v2elsnlsff <dbl>,
#> #   v2eldommon <dbl>, v2ellocons <dbl>, v2ellocumul <dbl>, v2elloeldm <dbl>,
#> #   v2elloelsy <dbl>, v2elloseat <dbl>, v2ellostlg <dbl>, v2ellostsl <dbl>,
#> #   v2ellostsm <dbl>, v2ellostss <dbl>, v2ellosttm <dbl>, v2ellostts <dbl>,
#> #   v2ellovtlg <dbl>, v2ellovtsm <dbl>, v2ellovttm <dbl>, v2elparlel <dbl>,
#> #   v2elprescons <dbl>, v2elprescumul <dbl>, v2elthresh <dbl>, v2eltvrig <dbl>,
#> #   v2elvotlrg <dbl>, v2elvotsml <dbl>, v2pscnslnl <dbl>, v2exaphogp <dbl>,
#> #   v2exaphos <dbl>, v2ex_elechog <dbl>, v2ex_elechos <dbl>, v2lgelecup <dbl>,
#> #   v2lgello <dbl>, v2lginello <dbl>, v2lginelup <dbl>, v3eldirepr <dbl>,
#> #   v3eldireuc <dbl>, v3eltvriguc <dbl>, v3elupseat <dbl>, v3elupstsl <dbl>,
#> #   v3elupstsm <dbl>, v3elupvtlg <dbl>, v3elupvtsm <dbl>, v3elloelsy <dbl>,
#> #   v3elloseat <dbl>, v3ellostlg <dbl>, v3ellostsl <dbl>, v3ellostsm <dbl>,
#> #   v3ellostss <dbl>, v3ellovtlg <dbl>, v3ellovtsm <dbl>, v3elparlel <dbl>,
#> #   v3eltvrig <dbl>, v3elvotlrg <dbl>, v3elvotsml <dbl>, v3lgelecup <dbl>,
#> #   v3lgello <dbl>, v3lginello <dbl>, v3lginelup <dbl>, v3ellocelc <dbl>,
#> #   v3elsrgel <dbl>, v2x_ex_direlect <dbl>, v2xex_elecreg <dbl>,
#> #   v2xlg_elecreg <dbl>, v2x_hosabort <dbl>, v2x_hosinter <dbl>,
#> #   v2x_legabort <dbl>, v2xel_elecparl <dbl>, v2xel_elecpres <dbl>,
#> #   v2xex_elecleg <dbl>, v2elsnlfc_0 <dbl>, v2elsnlfc_1 <dbl>,
#> #   v2elsnlfc_10 <dbl>, v2elsnlfc_11 <dbl>, v2elsnlfc_12 <dbl>,
#> #   v2elsnlfc_13 <dbl>, v2elsnlfc_14 <dbl>, v2elsnlfc_15 <dbl>,
#> #   v2elsnlfc_16 <dbl>, v2elsnlfc_17 <dbl>, v2elsnlfc_18 <dbl>, ...