Extracts parts of the V-Dem dataset by passing a filter expression to vdem_codebook

Extracts variables in the V-Dem dataset (version 9), by using a filter expression that is used to extract variable names from the V-Dem coddebook (vdem_codebook)

extract_vdem_filter(..., include_uncertainty = TRUE,
  include_sd = FALSE, include_osp = FALSE, include_ord = FALSE,
  include_nr = FALSE, include_external = FALSE)

Arguments

...
A sequence of filter expressions, separated by commas, suitable for use with dplyr::filter, operating on the column names of vdem_codebook. See examples.

include_uncertainty
Whether to return the uncertainty measurements for the selected variables (codelow and codehigh suffixes). Default is TRUE.

include_sd
Whether to return the standard deviation for the selected variables (sd suffix). Default is FALSE.

include_osp
Whether to return variables available in the original scale (osp suffix, linearized original scale posterior prediction). Default is FALSE.

include_ord
Whether to return variables available in an ordinal scale (ord suffix, measurement model estimates of original scale value). Default is FALSE.

include_nr
Whether to return the number of coders for the variable (nr suffix). Default is FALSE.

include_external
Whether to return variables marked "external", i.e., coming from an external source rather than V-Dem (e_ prefix). Default is FALSE.

Value

A tibble with the V-Dem country identifiers, other state system identifiers, plus the desired variables.

Examples

# All in section 1 of the codebook except component indexes
extract_vdem_filter(section == 1, !grepl("component index", label))

#> # A tibble: 26,884 x 32
#>    vdem_country_na~ vdem_country_te~ vdem_country_id  year extended_countr~
#>    <chr>            <chr>                      <dbl> <dbl> <chr>           
#>  1 Afghanistan      AFG                           36  1789 Afghanistan     
#>  2 Afghanistan      AFG                           36  1790 Afghanistan     
#>  3 Afghanistan      AFG                           36  1791 Afghanistan     
#>  4 Afghanistan      AFG                           36  1792 Afghanistan     
#>  5 Afghanistan      AFG                           36  1793 Afghanistan     
#>  6 Afghanistan      AFG                           36  1794 Afghanistan     
#>  7 Afghanistan      AFG                           36  1795 Afghanistan     
#>  8 Afghanistan      AFG                           36  1796 Afghanistan     
#>  9 Afghanistan      AFG                           36  1797 Afghanistan     
#> 10 Afghanistan      AFG                           36  1798 Afghanistan     
#> # ... with 26,874 more rows, and 27 more variables: GWn <dbl>,
#> #   historical_date <date>, project <dbl>, historical <int>, histname <chr>,
#> #   codingstart <dbl>, codingend <dbl>, codingstart_contemp <dbl>,
#> #   codingend_contemp <dbl>, codingstart_hist <dbl>, codingend_hist <dbl>,
#> #   gapstart1 <dbl>, gapstart2 <dbl>, gapstart3 <dbl>, gapend1 <dbl>,
#> #   gapend2 <dbl>, gapend3 <dbl>, vdem_cown <int>, GW_startdate <date>,
#> #   GW_enddate <date>, GWc <chr>, extended_region <chr>,
#> #   extended_continent <chr>, microstate <lgl>, lat <dbl>, lon <dbl>,
#> #   in_GW_system <lgl>

# All variables inclued in release 2 of the dataset
extract_vdem_filter(grepl("2(,|$)", data_release))

#> # A tibble: 26,884 x 35
#>    vdem_country_na~ vdem_country_te~ vdem_country_id  year extended_countr~
#>    <chr>            <chr>                      <dbl> <dbl> <chr>           
#>  1 Afghanistan      AFG                           36  1789 Afghanistan     
#>  2 Afghanistan      AFG                           36  1790 Afghanistan     
#>  3 Afghanistan      AFG                           36  1791 Afghanistan     
#>  4 Afghanistan      AFG                           36  1792 Afghanistan     
#>  5 Afghanistan      AFG                           36  1793 Afghanistan     
#>  6 Afghanistan      AFG                           36  1794 Afghanistan     
#>  7 Afghanistan      AFG                           36  1795 Afghanistan     
#>  8 Afghanistan      AFG                           36  1796 Afghanistan     
#>  9 Afghanistan      AFG                           36  1797 Afghanistan     
#> 10 Afghanistan      AFG                           36  1798 Afghanistan     
#> # ... with 26,874 more rows, and 30 more variables: GWn <dbl>,
#> #   historical_date <date>, project <dbl>, historical <int>, histname <chr>,
#> #   codingstart <dbl>, codingend <dbl>, codingstart_contemp <dbl>,
#> #   codingend_contemp <dbl>, codingstart_hist <dbl>, codingend_hist <dbl>,
#> #   gapstart1 <dbl>, gapstart2 <dbl>, gapstart3 <dbl>, gapend1 <dbl>,
#> #   gapend2 <dbl>, gapend3 <dbl>, vdem_cown <int>, v2x_libdem <dbl>,
#> #   v2x_libdem_codelow <dbl>, v2x_libdem_codehigh <dbl>, GW_startdate <date>,
#> #   GW_enddate <date>, GWc <chr>, extended_region <chr>,
#> #   extended_continent <chr>, microstate <lgl>, lat <dbl>, lon <dbl>,
#> #   in_GW_system <lgl>