V-Dem dataset (version 9)

The V-Dem dataset (version 9), including external variables. This documentation only provides variable names and labels for the country-year identifiers included. The rest of the codebook is available at https://www.v-dem.net/en/data/data-version-9/, and in this package as vdem_codebook.

VDem_plus

Format

An object of class tbl_df (inherits from tbl, data.frame) with 26884 rows and 3900 columns.

Source

Coppedge, Michael, John Gerring, Carl Henrik Knutsen, Staffan I. Lindberg, Jan Teorell, David Altman, Michael Bernhard, M. Steven Fish, Adam Glynn, Allen Hicken, Anna Lührmann, Kyle L. Marquardt, Kelly McMann, Pamela Paxton, Daniel Pemstein, Brigitte Seim, Rachel Sigman, Svend-Erik Skaaning, Jeffrey Staton, Agnes Cornell, Lisa Gastaldi, Haakon Gjerløw, Valeriya Mechkov, Johannes von Römer, Aksel Sundtröm, Eitan Tzelgov, Luca Uberti, Yi-ting Wang, Tore Wig, and Daniel Ziblatt. 2019. "V-Dem Codebook v9" Varieties of Democracy (V-Dem) Project. Available at: https://www.v-dem.net/en/data/data-version-9/

Identifiers

vdem_country_name

The original country name in the dataset. There are 178 distinct countries in version 9. According to the codebook, "a V-Dem country is a political unit enjoying at least some degree of functional and/or formal sovereignty."

vdem_country_id

Unique numerical country id for each country.

vdem_country_text_id

Unique abbreviation for each country.

year

Year, from 1789-2018.

extended_country_name

The standard country according to the Gleditsch and Ward system of states (Gleditsch and Ward 1999), or a standardized country name. Note that West Bank, and Gaza are two separate coding units in vdem_country_name but all map to "Palestine, State of" here. Column added by Xavier Marquez to the original V-Dem dataset.

GWn

Gleditsch and Ward's numeric country code, from the Gleditsch and Ward list of independent states. This differs from the V-Dem COW code in a few instances, and it does not exist in for some historical polities. For details, see Gleditsch and Ward 1999. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html. Column added by Xavier Marquez to the original V-Dem dataset.

historical_date

"This variable is included in the V-Dem Country Date dataset. The default date is December 31st, as in 2018-12-31, referring to the time span from 01-01 to 12-31 in a respective year. Additionally, specific changes, such as the appointment of a Head of State, are coded on the specific date within a certain year. Thus, a code can change within a year, and will be reflected in the 12-31 date."

codingstart

"V-Dem country coding starts in 1789, or from when a country first enjoyed at least some degree of functional and/or formal sovereignty. For detailed information, please see the V-Dem Country Coding Units document."

codingstart_contemp

This variable indicates the coding start for the countries coded by Contemporary V-Dem.

codingstart_hist

This variable indicates the coding start for the countries coded by Historical V-Dem.

gapstart

"Time periods when a country does not fulfill V-Dem’s coding period criteria are not coded. The date that indicates the gap start is the last date coded before the gap. For more details about V-Dem country coding periods, please see the V-Dem Country Coding Units document."

gapend

"The periods of when a country does not fulfill V-Dem’s coding period macriteria are not coded. The date that indicates the gap end is the first date coded after the gap. For more details about V-Dem country coding periods, please see the V-Dem Country Coding Units document."

codingend

"V-Dem country coding ends in 2018, or from when a country formally stopped enjoying at least some degree of functional and/or formal sovereignty. For detailed information, please see the V-Dem Country Coding Units document."

codingend_contemp

"This variable indicates when the coding ends for countries coded by the Contemporary V-Dem project."

codingend_hist

"This variable indicates when the coding ends for countries coded by the Historical V-Dem Project."

project

"This variable indicates which V-Dem project coded that country-year: Contemporary V-Dem (0), Historical V-Dem (1), or both (2)."

historical

"This variable indicates if the Historical V-Dem project coded a country at any time: No (0), Yes (1)."

vdem_cown

The Correlates of War numeric country code (2016). Note that this differs from Gleditsch and Ward's numeric country code in a few cases.

cown

The Correlates of War numeric country code (from the 2016 version of the COW state list). Note that this differs from Gleditsch and Ward's numeric country code in a few cases. Column added by Xavier Marquez to the original V-Dem dataset.

GWc

Gleditsch and Ward's alphabetic country code, from the Gleditsch and Ward list of independent states. Missing for some historical polities and some states with contested sovereignty. For details, see Gleditsch and Ward 1999. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html. Column added by Xavier Marquez to the original V-Dem dataset.

extended_region

Region. Column added by Xavier Marquez to the original V-Dem dataset. You may wish to use the e_region* set of variables instead.

extended_continent

Continent. Column added by Xavier Marquez to the original V-Dem dataset.

GW_startdate

The entry date of the state into the international system (usually the date of independence, though there can be more than one), according to Gleditsch and Ward, or 1 January 1816, whichever is later. Column added by Xavier Marquez to the original V-Dem dataset.

GW_enddate

The exit date of the state from the international system (usually the date the state lost its independence, though there can be more than one), according to Gleditsch and Ward, or NA if the country is still in the international system as of the time the dataset was compiled. Column added by Xavier Marquez to the original V-Dem dataset.

microstate

Whether the state is a microstate, according to Gleditsch. His tentative list of microstates is available at http://privatewww.essex.ac.uk/~ksg/statelist.html. Column added by Xavier Marquez to the original V-Dem dataset.

lat

The rough latitude of the state. Column added by Xavier Marquez to the original V-Dem dataset.

lon

The rough longitude of the state. Column added by Xavier Marquez to the original V-Dem dataset.

in_GW_system

Whether the state is "in system" (that is, is independent and sovereign), according to Gleditsch and Ward, for this particular date. Column added by Xavier Marquez to the original V-Dem dataset. You may instead wish to use one of the `v2sv*`` columns provided in the original V-Dem dataset.

References

Coppedge, Michael, John Gerring, Carl Henrik Knutsen, Staffan I. Lindberg, Jan Teorell, David Altman, Michael Bernhard, M. Steven Fish, Adam Glynn, Allen Hicken, Anna Lührmann, Kyle L. Marquardt, Kelly McMann, Pamela Paxton, Daniel Pemstein, Brigitte Seim, Rachel Sigman, Svend-Erik Skaaning, Jeffrey Staton, Steven Wilson, Agnes Cornell, Lisa Gastaldi, Haakon Gjerløw, Nina Ilchenko, Joshua Krusell, Valeriya Mechkova, Juraj Medzihorsky, Josefine Pernes, Johannes von Römer, Natalia Stepanova, Aksel Sundström, Eitan Tzelgov, Yi-ting Wang, Tore Wig, and Daniel Ziblatt. 2019. "V-Dem Country-Year/Country-Date Dataset v9" Varieties of Democracy (V-Dem) Project. Available at https://www.v-dem.net/en/data/data-version-9/

Pemstein, Daniel, Kyle L. Marquardt, Eitan Tzelgov, Yi-ting Wang, Juraj Medzihorsky, Joshua Krusell, Farhad Miri, and Johannes von Römer. 2019. “The V-Dem Measurement Model: Latent Variable Analysis for Cross-National and Cross-Temporal Expert-Coded Data”. V-Dem Working Paper No. 21. 4th edition. University of Gothenburg: Varieties of Democracy Institute.

Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html