Extracts parts of the V-Dem dataset by section number, name pattern, or label pattern

Extracts variables in the V-Dem dataset (version 7.1), according to the section number in the codebook or a regular expression pattern matching the name or the label.

extract_vdem(section_number = NULL, name_pattern = NULL,
  label_pattern = NULL, include_uncertainty = TRUE, include_sd = FALSE,
  include_osp = FALSE, include_ord = FALSE, include_nr = FALSE,
  include_external = FALSE)

Arguments

section_number
A section number (or vector of section numbers) in the V-dem codebook. Must be an integer or an integer vector (e.g., 1:2, or 28).

name_pattern
A character pattern, interpreted as a regular expression, to search over the V-dem variable names, e.g., "v2x" extracts all variables including the string "v2x" in their names.

label_pattern
A character pattern, interpreted as a regular expression, to search over the V-dem labels, e.g., "corrupt" extracts all variables that mention "corrupt" in their label.

include_uncertainty
Whether to return the uncertainty measurements for the selected variables (codelow and codehigh suffixes). Default is TRUE.

include_sd
Whether to return the standard deviation for the selected variables (sd suffix). Default is FALSE.

include_osp
Whether to return variables available in the original scale (osp suffix, linearized original scale posterior prediction). Default is FALSE.

include_ord
Whether to return variables available in an ordinal scale (ord suffix, measurement model estimates of original scale value). Default is FALSE.

include_nr
Whether to return the number of coders for the variable (nr suffix). Default is FALSE.

include_external
Whether to return variables marked "external", i.e., coming from an external source rather than V-Dem (e_ prefix). Default is FALSE.

Value

A tibble with the V-Dem country identifiers, other state system identifiers, plus the desired variables.

Examples

# Main democracy indexes
extract_vdem(section_number = 1)

#> # A tibble: 17,604 x 55
#>    vdem_country_name vdem_country_id vdem_country_text_id  year
#>                <chr>           <dbl>                <chr> <dbl>
#>  1       Afghanistan              36                  AFG  1900
#>  2       Afghanistan              36                  AFG  1901
#>  3       Afghanistan              36                  AFG  1902
#>  4       Afghanistan              36                  AFG  1903
#>  5       Afghanistan              36                  AFG  1904
#>  6       Afghanistan              36                  AFG  1905
#>  7       Afghanistan              36                  AFG  1906
#>  8       Afghanistan              36                  AFG  1907
#>  9       Afghanistan              36                  AFG  1908
#> 10       Afghanistan              36                  AFG  1909
#> # ... with 17,594 more rows, and 51 more variables:
#> #   extended_country_name <chr>, GWn <dbl>, historical_date <date>,
#> #   codingstart <date>, gapstart <date>, gapend <date>, codingend <date>,
#> #   vdem_cown <dbl>, cown <int>, v2x_polyarchy <dbl>, v2x_api <dbl>,
#> #   v2x_mpi <dbl>, v2x_libdem <dbl>, v2x_liberal <dbl>, v2x_partipdem <dbl>,
#> #   v2x_partip <dbl>, v2x_delibdem <dbl>, v2xdl_delib <dbl>, v2x_egaldem <dbl>,
#> #   v2x_egal <dbl>, v2x_polyarchy_codelow <dbl>, v2x_api_codelow <dbl>,
#> #   v2x_mpi_codelow <dbl>, v2x_libdem_codelow <dbl>, v2x_liberal_codelow <dbl>,
#> #   v2x_partipdem_codelow <dbl>, v2x_partip_codelow <dbl>,
#> #   v2x_delibdem_codelow <dbl>, v2xdl_delib_codelow <dbl>,
#> #   v2x_egaldem_codelow <dbl>, v2x_egal_codelow <dbl>,
#> #   v2x_polyarchy_codehigh <dbl>, v2x_api_codehigh <dbl>,
#> #   v2x_mpi_codehigh <dbl>, v2x_libdem_codehigh <dbl>,
#> #   v2x_liberal_codehigh <dbl>, v2x_partipdem_codehigh <dbl>,
#> #   v2x_partip_codehigh <dbl>, v2x_delibdem_codehigh <dbl>,
#> #   v2xdl_delib_codehigh <dbl>, v2x_egaldem_codehigh <dbl>,
#> #   v2x_egal_codehigh <dbl>, GW_startdate <date>, GW_enddate <date>, GWc <chr>,
#> #   extended_region <chr>, extended_continent <chr>, microstate <lgl>,
#> #   lat <dbl>, lon <dbl>, in_GW_system <lgl>

# Mid-level indexes, no measures of uncertainty
extract_vdem(section_number = 2, include_uncertainty = FALSE)

#> # A tibble: 17,604 x 68
#>    vdem_country_name vdem_country_id vdem_country_text_id  year
#>                <chr>           <dbl>                <chr> <dbl>
#>  1       Afghanistan              36                  AFG  1900
#>  2       Afghanistan              36                  AFG  1901
#>  3       Afghanistan              36                  AFG  1902
#>  4       Afghanistan              36                  AFG  1903
#>  5       Afghanistan              36                  AFG  1904
#>  6       Afghanistan              36                  AFG  1905
#>  7       Afghanistan              36                  AFG  1906
#>  8       Afghanistan              36                  AFG  1907
#>  9       Afghanistan              36                  AFG  1908
#> 10       Afghanistan              36                  AFG  1909
#> # ... with 17,594 more rows, and 64 more variables:
#> #   extended_country_name <chr>, GWn <dbl>, historical_date <date>,
#> #   codingstart <date>, gapstart <date>, gapend <date>, codingend <date>,
#> #   vdem_cown <dbl>, cown <int>, v2x_EDcomp_thick <dbl>,
#> #   v2x_frassoc_thick <dbl>, v2x_freexp_thick <dbl>, v2x_freexp <dbl>,
#> #   v2xme_altinf <dbl>, v2x_suffr <dbl>, v2xel_frefair <dbl>,
#> #   v2x_elecoff <dbl>, v2xcl_rol <dbl>, v2x_jucon <dbl>, v2xlg_legcon <dbl>,
#> #   v2x_cspart <dbl>, v2xdd_dd <dbl>, v2xdd_cic <dbl>, v2xdd_toc <dbl>,
#> #   v2xel_locelec <dbl>, v2xel_regelec <dbl>, v2xeg_eqprotec <dbl>,
#> #   v2xeg_eqaccess <dbl>, v2xeg_eqdr <dbl>, v2xcs_ccsi <dbl>,
#> #   v2xps_party <dbl>, v2x_gender <dbl>, v2x_gencl <dbl>, v2x_gencs <dbl>,
#> #   v2x_genpp <dbl>, v2x_elecreg <dbl>, v2xex_elecreg <dbl>,
#> #   v2xlg_elecreg <dbl>, v2xel_elecparl <dbl>, v2xlg_leginter <dbl>,
#> #   v2xel_elecpres <dbl>, v2x_hosinter <dbl>, v2x_corr <dbl>,
#> #   v2x_pubcorr <dbl>, v2x_execorr <dbl>, v2x_divparctrl <dbl>,
#> #   v2x_feduni <dbl>, v2x_civlib <dbl>, v2x_clpriv <dbl>, v2x_clpol <dbl>,
#> #   v2x_clphy <dbl>, v2x_accountability <dbl>, v2x_veracc <dbl>,
#> #   v2x_horacc <dbl>, v2x_diagacc <dbl>, GW_startdate <date>,
#> #   GW_enddate <date>, GWc <chr>, extended_region <chr>,
#> #   extended_continent <chr>, microstate <lgl>, lat <dbl>, lon <dbl>,
#> #   in_GW_system <lgl>

# All indexes about corruption
extract_vdem(label_pattern = "corrupt")

#> # A tibble: 17,604 x 46
#>    vdem_country_name vdem_country_id vdem_country_text_id  year
#>                <chr>           <dbl>                <chr> <dbl>
#>  1       Afghanistan              36                  AFG  1900
#>  2       Afghanistan              36                  AFG  1901
#>  3       Afghanistan              36                  AFG  1902
#>  4       Afghanistan              36                  AFG  1903
#>  5       Afghanistan              36                  AFG  1904
#>  6       Afghanistan              36                  AFG  1905
#>  7       Afghanistan              36                  AFG  1906
#>  8       Afghanistan              36                  AFG  1907
#>  9       Afghanistan              36                  AFG  1908
#> 10       Afghanistan              36                  AFG  1909
#> # ... with 17,594 more rows, and 42 more variables:
#> #   extended_country_name <chr>, GWn <dbl>, historical_date <date>,
#> #   codingstart <date>, gapstart <date>, gapend <date>, codingend <date>,
#> #   vdem_cown <dbl>, cown <int>, v2x_corr <dbl>, v2x_pubcorr <dbl>,
#> #   v2x_execorr <dbl>, v2exbribe <dbl>, v2excrptps <dbl>, v2lgcrrpt <dbl>,
#> #   v2jucorrdc <dbl>, v2mecorrpt <dbl>, v2x_corr_codelow <dbl>,
#> #   v2x_pubcorr_codelow <dbl>, v2x_execorr_codelow <dbl>,
#> #   v2exbribe_codelow <dbl>, v2excrptps_codelow <dbl>, v2lgcrrpt_codelow <dbl>,
#> #   v2jucorrdc_codelow <dbl>, v2mecorrpt_codelow <dbl>,
#> #   v2x_corr_codehigh <dbl>, v2x_pubcorr_codehigh <dbl>,
#> #   v2x_execorr_codehigh <dbl>, v2exbribe_codehigh <dbl>,
#> #   v2excrptps_codehigh <dbl>, v2lgcrrpt_codehigh <dbl>,
#> #   v2jucorrdc_codehigh <dbl>, v2mecorrpt_codehigh <dbl>, GW_startdate <date>,
#> #   GW_enddate <date>, GWc <chr>, extended_region <chr>,
#> #   extended_continent <chr>, microstate <lgl>, lat <dbl>, lon <dbl>,
#> #   in_GW_system <lgl>

# All indexes about elections, include external indexes
extract_vdem(label_pattern = "elect", include_external = TRUE)

#> # A tibble: 17,604 x 181
#>    vdem_country_name vdem_country_id vdem_country_text_id  year
#>                <chr>           <dbl>                <chr> <dbl>
#>  1       Afghanistan              36                  AFG  1900
#>  2       Afghanistan              36                  AFG  1901
#>  3       Afghanistan              36                  AFG  1902
#>  4       Afghanistan              36                  AFG  1903
#>  5       Afghanistan              36                  AFG  1904
#>  6       Afghanistan              36                  AFG  1905
#>  7       Afghanistan              36                  AFG  1906
#>  8       Afghanistan              36                  AFG  1907
#>  9       Afghanistan              36                  AFG  1908
#> 10       Afghanistan              36                  AFG  1909
#> # ... with 17,594 more rows, and 177 more variables:
#> #   extended_country_name <chr>, GWn <dbl>, historical_date <date>,
#> #   codingstart <date>, gapstart <date>, gapend <date>, codingend <date>,
#> #   vdem_cown <dbl>, cown <int>, v2xel_frefair <dbl>, v2xex_elecreg <dbl>,
#> #   v2xlg_elecreg <dbl>, v2xel_elecparl <dbl>, v2xel_elecpres <dbl>,
#> #   v2x_hosinter <dbl>, v2x_hosabort <dbl>, v2x_legabort <dbl>,
#> #   v2elmulpar_ex <dbl>, v2elmulpar_leg <dbl>, v2elrgstry_ex <dbl>,
#> #   v2elrgstry_leg <dbl>, v2elvotbuy_ex <dbl>, v2elvotbuy_leg <dbl>,
#> #   v2elirreg_ex <dbl>, v2elirreg_leg <dbl>, v2elintim_ex <dbl>,
#> #   v2elintim_leg <dbl>, v2elpeace <dbl>, v2elpeace_ex <dbl>,
#> #   v2elpeace_leg <dbl>, v2elboycot_ex <dbl>, v2elboycot_leg <dbl>,
#> #   v2elfrcamp_ex <dbl>, v2elfrcamp_leg <dbl>, v2elpdcamp_ex <dbl>,
#> #   v2elpdcamp_leg <dbl>, v2elpaidig_ex <dbl>, v2elpaidig_leg <dbl>,
#> #   v2eldommon <dbl>, v2eldommon_ex <dbl>, v2eldommon_leg <dbl>,
#> #   v2elintmon_ex <dbl>, v2elintmon_leg <dbl>, v2elmonden_ex <dbl>,
#> #   v2elmonden_leg <dbl>, v2elwestmon_ex <dbl>, v2elwestmon_leg <dbl>,
#> #   v2elfrfair_ex <dbl>, v2elfrfair_leg <dbl>, v2elvaptrn_ex <dbl>,
#> #   v2elvaptrn_leg <dbl>, v2eltrnout_ex <dbl>, v2eltrnout_leg <dbl>,
#> #   v2elaccept_ex <dbl>, v2elaccept_leg <dbl>, v2elasmoff_ex <dbl>,
#> #   v2elasmoff_leg <dbl>, v2elprescumul <dbl>, v2elprescons <dbl>,
#> #   v2elvotlrg <dbl>, v2elvotsml <dbl>, v2ellocumul <dbl>, v2ellocons <dbl>,
#> #   v2elparlel <dbl>, v2elloelsy <dbl>, v2elloeldm <dbl>, v2elthresh <dbl>,
#> #   v2ellovtlg <dbl>, v2ellovtsm <dbl>, v2ellostlg <dbl>, v2elloseat <dbl>,
#> #   v2ellostsl <dbl>, v2ellostsm <dbl>, v2ellostss <dbl>, v2eltvrig <dbl>,
#> #   v2elsrgel <dbl>, v2ellocelc <dbl>, v2elffelr <dbl>, v2elffelrbin <dbl>,
#> #   v2elsnlsff <dbl>, v2pscnslnl <dbl>, v2ex_elechos <dbl>, v2exaphos <dbl>,
#> #   v2ex_elechog <dbl>, v2exaphogp <dbl>, v2lgello <dbl>, v2ex_elecleg <dbl>,
#> #   v2lgelecup <dbl>, e_v2xel_frefair_3C <dbl>, e_v2xel_frefair_4C <dbl>,
#> #   e_v2xel_frefair_5C <dbl>, e_electedper <dbl>, e_SR11CHEAT <dbl>,
#> #   e_SR13VIOL <dbl>, e_SR12CAP <dbl>, e_competition <dbl>,
#> #   e_multiparty_elections <dbl>, e_exselec <dbl>, e_legselec <dbl>,
#> #   e_ellonmpe <dbl>, ...

# All civil liberties indexes, include original scale prediction
extract_vdem(name_pattern = "^v2cl", include_osp = TRUE)

#> # A tibble: 17,604 x 195
#>    vdem_country_name vdem_country_id vdem_country_text_id  year
#>                <chr>           <dbl>                <chr> <dbl>
#>  1       Afghanistan              36                  AFG  1900
#>  2       Afghanistan              36                  AFG  1901
#>  3       Afghanistan              36                  AFG  1902
#>  4       Afghanistan              36                  AFG  1903
#>  5       Afghanistan              36                  AFG  1904
#>  6       Afghanistan              36                  AFG  1905
#>  7       Afghanistan              36                  AFG  1906
#>  8       Afghanistan              36                  AFG  1907
#>  9       Afghanistan              36                  AFG  1908
#> 10       Afghanistan              36                  AFG  1909
#> # ... with 17,594 more rows, and 191 more variables:
#> #   extended_country_name <chr>, GWn <dbl>, historical_date <date>,
#> #   codingstart <date>, gapstart <date>, gapend <date>, codingend <date>,
#> #   vdem_cown <dbl>, cown <int>, v2clacfree <dbl>, v2clacfree_osp <dbl>,
#> #   v2clrelig <dbl>, v2clrelig_osp <dbl>, v2cltort <dbl>, v2cltort_osp <dbl>,
#> #   v2clkill <dbl>, v2clkill_osp <dbl>, v2cltrnslw <dbl>, v2cltrnslw_osp <dbl>,
#> #   v2clrspct <dbl>, v2clrspct_osp <dbl>, v2clfmove <dbl>, v2clfmove_osp <dbl>,
#> #   v2cldmovem <dbl>, v2cldmovem_osp <dbl>, v2cldmovew <dbl>,
#> #   v2cldmovew_osp <dbl>, v2cldiscm <dbl>, v2cldiscm_osp <dbl>,
#> #   v2cldiscw <dbl>, v2cldiscw_osp <dbl>, v2clslavem <dbl>,
#> #   v2clslavem_osp <dbl>, v2clslavef <dbl>, v2clslavef_osp <dbl>,
#> #   v2clstown <dbl>, v2clstown_osp <dbl>, v2clprptym <dbl>,
#> #   v2clprptym_osp <dbl>, v2clprptyw <dbl>, v2clprptyw_osp <dbl>,
#> #   v2clacjstm <dbl>, v2clacjstm_osp <dbl>, v2clacjstw <dbl>,
#> #   v2clacjstw_osp <dbl>, v2clacjust <dbl>, v2clacjust_osp <dbl>,
#> #   v2clsocgrp <dbl>, v2clsocgrp_osp <dbl>, v2clrgunev <dbl>,
#> #   v2clrgunev_osp <dbl>, v2clsnlpct <dbl>, v2clrgstch_0 <dbl>,
#> #   v2clrgstch_1 <dbl>, v2clrgstch_10 <dbl>, v2clrgstch_11 <dbl>,
#> #   v2clrgstch_12 <dbl>, v2clrgstch_13 <dbl>, v2clrgstch_14 <dbl>,
#> #   v2clrgstch_15 <dbl>, v2clrgstch_16 <dbl>, v2clrgstch_17 <dbl>,
#> #   v2clrgstch_18 <dbl>, v2clrgstch_19 <dbl>, v2clrgstch_2 <dbl>,
#> #   v2clrgstch_20 <dbl>, v2clrgstch_21 <dbl>, v2clrgstch_3 <dbl>,
#> #   v2clrgstch_4 <dbl>, v2clrgstch_5 <dbl>, v2clrgstch_6 <dbl>,
#> #   v2clrgstch_7 <dbl>, v2clrgstch_8 <dbl>, v2clrgstch_9 <dbl>,
#> #   v2clrgwkch_0 <dbl>, v2clrgwkch_1 <dbl>, v2clrgwkch_10 <dbl>,
#> #   v2clrgwkch_11 <dbl>, v2clrgwkch_12 <dbl>, v2clrgwkch_13 <dbl>,
#> #   v2clrgwkch_14 <dbl>, v2clrgwkch_15 <dbl>, v2clrgwkch_16 <dbl>,
#> #   v2clrgwkch_17 <dbl>, v2clrgwkch_18 <dbl>, v2clrgwkch_19 <dbl>,
#> #   v2clrgwkch_2 <dbl>, v2clrgwkch_20 <dbl>, v2clrgwkch_21 <dbl>,
#> #   v2clrgwkch_3 <dbl>, v2clrgwkch_4 <dbl>, v2clrgwkch_5 <dbl>,
#> #   v2clrgwkch_6 <dbl>, v2clrgwkch_7 <dbl>, v2clrgwkch_8 <dbl>,
#> #   v2clrgwkch_9 <dbl>, v2clacfree_codelow <dbl>, v2clacfree_osp_codelow <dbl>,
#> #   v2clrelig_codelow <dbl>, v2clrelig_osp_codelow <dbl>, ...

# Combine section and label patterns
extract_vdem(section_number = 1, label_pattern = "elect", include_external = TRUE)

#> # A tibble: 17,604 x 214
#>    vdem_country_name vdem_country_id vdem_country_text_id  year
#>                <chr>           <dbl>                <chr> <dbl>
#>  1       Afghanistan              36                  AFG  1900
#>  2       Afghanistan              36                  AFG  1901
#>  3       Afghanistan              36                  AFG  1902
#>  4       Afghanistan              36                  AFG  1903
#>  5       Afghanistan              36                  AFG  1904
#>  6       Afghanistan              36                  AFG  1905
#>  7       Afghanistan              36                  AFG  1906
#>  8       Afghanistan              36                  AFG  1907
#>  9       Afghanistan              36                  AFG  1908
#> 10       Afghanistan              36                  AFG  1909
#> # ... with 17,594 more rows, and 210 more variables:
#> #   extended_country_name <chr>, GWn <dbl>, historical_date <date>,
#> #   codingstart <date>, gapstart <date>, gapend <date>, codingend <date>,
#> #   vdem_cown <dbl>, cown <int>, v2x_polyarchy <dbl>, v2x_api <dbl>,
#> #   v2x_mpi <dbl>, v2x_libdem <dbl>, v2x_liberal <dbl>, v2x_partipdem <dbl>,
#> #   v2x_partip <dbl>, v2x_delibdem <dbl>, v2xdl_delib <dbl>, v2x_egaldem <dbl>,
#> #   v2x_egal <dbl>, v2x_polyarchy_codelow <dbl>, v2x_api_codelow <dbl>,
#> #   v2x_mpi_codelow <dbl>, v2x_libdem_codelow <dbl>, v2x_liberal_codelow <dbl>,
#> #   v2x_partipdem_codelow <dbl>, v2x_partip_codelow <dbl>,
#> #   v2x_delibdem_codelow <dbl>, v2xdl_delib_codelow <dbl>,
#> #   v2x_egaldem_codelow <dbl>, v2x_egal_codelow <dbl>,
#> #   v2x_polyarchy_codehigh <dbl>, v2x_api_codehigh <dbl>,
#> #   v2x_mpi_codehigh <dbl>, v2x_libdem_codehigh <dbl>,
#> #   v2x_liberal_codehigh <dbl>, v2x_partipdem_codehigh <dbl>,
#> #   v2x_partip_codehigh <dbl>, v2x_delibdem_codehigh <dbl>,
#> #   v2xdl_delib_codehigh <dbl>, v2x_egaldem_codehigh <dbl>,
#> #   v2x_egal_codehigh <dbl>, v2xel_frefair <dbl>, v2xex_elecreg <dbl>,
#> #   v2xlg_elecreg <dbl>, v2xel_elecparl <dbl>, v2xel_elecpres <dbl>,
#> #   v2x_hosinter <dbl>, v2x_hosabort <dbl>, v2x_legabort <dbl>,
#> #   v2elmulpar_ex <dbl>, v2elmulpar_leg <dbl>, v2elrgstry_ex <dbl>,
#> #   v2elrgstry_leg <dbl>, v2elvotbuy_ex <dbl>, v2elvotbuy_leg <dbl>,
#> #   v2elirreg_ex <dbl>, v2elirreg_leg <dbl>, v2elintim_ex <dbl>,
#> #   v2elintim_leg <dbl>, v2elpeace <dbl>, v2elpeace_ex <dbl>,
#> #   v2elpeace_leg <dbl>, v2elboycot_ex <dbl>, v2elboycot_leg <dbl>,
#> #   v2elfrcamp_ex <dbl>, v2elfrcamp_leg <dbl>, v2elpdcamp_ex <dbl>,
#> #   v2elpdcamp_leg <dbl>, v2elpaidig_ex <dbl>, v2elpaidig_leg <dbl>,
#> #   v2eldommon <dbl>, v2eldommon_ex <dbl>, v2eldommon_leg <dbl>,
#> #   v2elintmon_ex <dbl>, v2elintmon_leg <dbl>, v2elmonden_ex <dbl>,
#> #   v2elmonden_leg <dbl>, v2elwestmon_ex <dbl>, v2elwestmon_leg <dbl>,
#> #   v2elfrfair_ex <dbl>, v2elfrfair_leg <dbl>, v2elvaptrn_ex <dbl>,
#> #   v2elvaptrn_leg <dbl>, v2eltrnout_ex <dbl>, v2eltrnout_leg <dbl>,
#> #   v2elaccept_ex <dbl>, v2elaccept_leg <dbl>, v2elasmoff_ex <dbl>,
#> #   v2elasmoff_leg <dbl>, v2elprescumul <dbl>, v2elprescons <dbl>,
#> #   v2elvotlrg <dbl>, v2elvotsml <dbl>, v2ellocumul <dbl>, v2ellocons <dbl>,
#> #   v2elparlel <dbl>, v2elloelsy <dbl>, v2elloeldm <dbl>, v2elthresh <dbl>, ...