Extracts parts of the V-Dem dataset by passing a filter expression to vdem_codebook

Extracts variables in the V-Dem dataset (version 7.1), by using a filter expression that is used to extract variable names from the V-Dem coddebook (vdem_codebook)

extract_vdem_filter(..., include_uncertainty = TRUE, include_sd = FALSE,
  include_osp = FALSE, include_ord = FALSE, include_nr = FALSE,
  include_external = FALSE)

Arguments

...
A sequence of filter expressions, separated by commas, suitable for use with dplyr::filter, operating on the column names of vdem_codebook. See examples.

include_uncertainty
Whether to return the uncertainty measurements for the selected variables (codelow and codehigh suffixes). Default is TRUE.

include_sd
Whether to return the standard deviation for the selected variables (sd suffix). Default is FALSE.

include_osp
Whether to return variables available in the original scale (osp suffix, linearized original scale posterior prediction). Default is FALSE.

include_ord
Whether to return variables available in an ordinal scale (ord suffix, measurement model estimates of original scale value). Default is FALSE.

include_nr
Whether to return the number of coders for the variable (nr suffix). Default is FALSE.

include_external
Whether to return variables marked "external", i.e., coming from an external source rather than V-Dem (e_ prefix). Default is FALSE.

Value

A tibble with the V-Dem country identifiers, other state system identifiers, plus the desired variables.

Examples

# All in section 1 of the codebook except component indexes
extract_vdem_filter(section == 1, !grepl("component index", label))

#> # A tibble: 17,604 x 43
#>    vdem_country_name vdem_country_id vdem_country_text_id  year
#>                <chr>           <dbl>                <chr> <dbl>
#>  1       Afghanistan              36                  AFG  1900
#>  2       Afghanistan              36                  AFG  1901
#>  3       Afghanistan              36                  AFG  1902
#>  4       Afghanistan              36                  AFG  1903
#>  5       Afghanistan              36                  AFG  1904
#>  6       Afghanistan              36                  AFG  1905
#>  7       Afghanistan              36                  AFG  1906
#>  8       Afghanistan              36                  AFG  1907
#>  9       Afghanistan              36                  AFG  1908
#> 10       Afghanistan              36                  AFG  1909
#> # ... with 17,594 more rows, and 39 more variables:
#> #   extended_country_name <chr>, GWn <dbl>, historical_date <date>,
#> #   codingstart <date>, gapstart <date>, gapend <date>, codingend <date>,
#> #   vdem_cown <dbl>, cown <int>, v2x_polyarchy <dbl>, v2x_api <dbl>,
#> #   v2x_mpi <dbl>, v2x_libdem <dbl>, v2x_partipdem <dbl>, v2x_delibdem <dbl>,
#> #   v2x_egaldem <dbl>, v2x_polyarchy_codelow <dbl>, v2x_api_codelow <dbl>,
#> #   v2x_mpi_codelow <dbl>, v2x_libdem_codelow <dbl>,
#> #   v2x_partipdem_codelow <dbl>, v2x_delibdem_codelow <dbl>,
#> #   v2x_egaldem_codelow <dbl>, v2x_polyarchy_codehigh <dbl>,
#> #   v2x_api_codehigh <dbl>, v2x_mpi_codehigh <dbl>, v2x_libdem_codehigh <dbl>,
#> #   v2x_partipdem_codehigh <dbl>, v2x_delibdem_codehigh <dbl>,
#> #   v2x_egaldem_codehigh <dbl>, GW_startdate <date>, GW_enddate <date>,
#> #   GWc <chr>, extended_region <chr>, extended_continent <chr>,
#> #   microstate <lgl>, lat <dbl>, lon <dbl>, in_GW_system <lgl>

# All variables inclued in release 2 of the dataset
extract_vdem_filter(grepl("2(,|$)", data_release))

#> # A tibble: 17,604 x 794
#>    vdem_country_name vdem_country_id vdem_country_text_id  year
#>                <chr>           <dbl>                <chr> <dbl>
#>  1       Afghanistan              36                  AFG  1900
#>  2       Afghanistan              36                  AFG  1901
#>  3       Afghanistan              36                  AFG  1902
#>  4       Afghanistan              36                  AFG  1903
#>  5       Afghanistan              36                  AFG  1904
#>  6       Afghanistan              36                  AFG  1905
#>  7       Afghanistan              36                  AFG  1906
#>  8       Afghanistan              36                  AFG  1907
#>  9       Afghanistan              36                  AFG  1908
#> 10       Afghanistan              36                  AFG  1909
#> # ... with 17,594 more rows, and 790 more variables:
#> #   extended_country_name <chr>, GWn <dbl>, historical_date <date>,
#> #   codingstart <date>, gapstart <date>, gapend <date>, codingend <date>,
#> #   vdem_cown <dbl>, cown <int>, v2x_polyarchy <dbl>, v2x_libdem <dbl>,
#> #   v2x_liberal <dbl>, v2x_partipdem <dbl>, v2x_partip <dbl>,
#> #   v2x_delibdem <dbl>, v2xdl_delib <dbl>, v2x_egaldem <dbl>, v2x_egal <dbl>,
#> #   v2x_frassoc_thick <dbl>, v2x_freexp <dbl>, v2xme_altinf <dbl>,
#> #   v2x_suffr <dbl>, v2xel_frefair <dbl>, v2x_elecoff <dbl>, v2xcl_rol <dbl>,
#> #   v2x_jucon <dbl>, v2xlg_legcon <dbl>, v2x_cspart <dbl>, v2xdd_dd <dbl>,
#> #   v2xel_locelec <dbl>, v2xel_regelec <dbl>, v2xcs_ccsi <dbl>,
#> #   v2xps_party <dbl>, v2x_elecreg <dbl>, v2xel_elecparl <dbl>,
#> #   v2xlg_leginter <dbl>, v2xel_elecpres <dbl>, v2x_hosinter <dbl>,
#> #   v2elrstrct <dbl>, v2eldonate <dbl>, v2elpubfin <dbl>, v2elcomvot <dbl>,
#> #   v2elsuffrage <dbl>, v2elembaut <dbl>, v2elembcap <dbl>, v2x_hosabort <dbl>,
#> #   v2x_legabort <dbl>, v2elmulpar <dbl>, v2elrgstry <dbl>, v2elage <dbl>,
#> #   v2elgvsuflvl <dbl>, v2elvotbuy <dbl>, v2elirreg <dbl>, v2elintim <dbl>,
#> #   v2elpeace <dbl>, v2elboycot <dbl>, v2elfrcamp <dbl>, v2elpdcamp <dbl>,
#> #   v2elpaidig <dbl>, v2eldommon <dbl>, v2elintmon <dbl>, v2elmonden <dbl>,
#> #   v2elfrfair <dbl>, v2elvaptrn <dbl>, v2eltvrexo <dbl>, v2elaccept <dbl>,
#> #   v2elasmoff <dbl>, v2elturnhos <dbl>, v2elwomcab <dbl>, v2elvotlrg <dbl>,
#> #   v2elvotsml <dbl>, v2elthresh <dbl>, v2ellovtlg <dbl>, v2ellovtsm <dbl>,
#> #   v2ellostlg <dbl>, v2elloseat <dbl>, v2ellostsl <dbl>, v2ellostsm <dbl>,
#> #   v2ellostss <dbl>, v2eltvrig <dbl>, v2elsrgel <dbl>, v2elrgpwr <dbl>,
#> #   v2ellocelc <dbl>, v2ellocpwr <dbl>, v2elsnlsff <dbl>, v2psparban <dbl>,
#> #   v2psbars <dbl>, v2psoppaut <dbl>, v2psorgs <dbl>, v2psprbrch <dbl>,
#> #   v2psprlnks <dbl>, v2psplats <dbl>, v2pscnslnl <dbl>, v2pscohesv <dbl>,
#> #   v2pscomprg <dbl>, v2psnatpar <dbl>, v2pssunpar <dbl>, v2ddlexpl <dbl>,
#> #   v2ddlevpl <dbl+lbl>, v2ddpartpl <dbl>, ...